LPS-57354 Move constants classes to its place
